mcpclient/requirements.txt
chainlit>=1.0.0
openai>=1.0.0
httpx>=0.25.0
httpx-sse>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
mcp>=1.0.0
